chatbot/requirements.txt
fastapi
uvicorn
tortoise-orm
asyncpg
python-dotenv
openai
twilio
python-multipart
requests